+++ docs/packages/design-system/designSystem.navigation.ts
@@ -1,4 +1,4 @@
-import type { PackageDocNavigation } from '@docs/packages/navigation'
+import type { PackageDocNavigation } from '@docs/packages/navigation.utils'
 
 export const DESIGN_SYSTEM_NAVIGATION: PackageDocNavigation = {
   link: 'pages/getting-started/installation',
@@ -10,7 +10,200 @@ export const DESIGN_SYSTEM_NAVIGATION: PackageDocNavigation = {
       items: [
         {
           text: 'Installation',
-          link: '/pages/getting-started/installation',
+          link: 'pages/getting-started/installation',
+        },
+      ],
+    },
+    {
+      text: 'Actions',
+      items: [
+        {
+          text: 'Button',
+          link: 'components/button/button',
+        },
+        {
+          text: 'Icon Button',
+          link: 'components/icon-button/icon-button',
+        },
+        {
+          text: 'Link',
+          link: 'components/link/link',
+        },
+        {
+          text: 'Menu Item',
+          link: 'components/menu-item/menu-item',
+        },
+        {
+          text: 'Dropdown Menu',
+          link: 'components/dropdown-menu/dropdown-menu',
+        },
+        {
+          text: 'Context Menu',
+          link: 'components/context-menu/context-menu',
+        },
+      ],
+    },
+    {
+      text: 'Forms and inputs',
+      items: [
+        {
+          text: 'Address Autocomplete',
+          link: 'components/address-autocomplete/address-autocomplete',
+        },
+        {
+          text: 'Autocomplete',
+          link: 'components/autocomplete/autocomplete',
+        },
+        {
+          text: 'Checkbox',
+          link: 'components/checkbox/checkbox',
+        },
+        {
+          text: 'Checkbox Group',
+          link: 'components/checkbox-group/checkbox-group',
+        },
+        {
+          text: 'Date Field',
+          link: 'components/date-field/date-field',
+        },
+        {
+          text: 'Date Picker',
+          link: 'components/date-picker/date-picker',
+        },
+        {
+          text: 'Date Range Field',
+          link: 'components/date-range-field/date-range-field',
+        },
+        {
+          text: 'Date Range Picker',
+          link: 'components/date-range-picker/date-range-picker',
+        },
+        {
+          text: 'Number Field',
+          link: 'components/number-field/number-field',
+        },
+        {
+          text: 'Phone Number Field',
+          link: 'components/phone-number-field/phone-number-field',
+        },
+        {
+          text: 'Radio Group',
+          link: 'components/radio-group/radio-group',
+        },
+        {
+          text: 'Select',
+          link: 'components/select/select',
+        },
+        {
+          text: 'Switch',
+          link: 'components/switch/switch',
+        },
+        {
+          text: 'Text Field',
+          link: 'components/text-field/text-field',
+        },
+        {
+          text: 'Textarea Field',
+          link: 'components/textarea-field/textarea-field',
+        },
+        {
+          text: 'Time Field',
+          link: 'components/time-field/time-field',
+        },
+      ],
+    },
+    {
+      text: 'Feedback',
+      items: [
+        {
+          text: 'Dialog',
+          link: 'components/dialog/dialog',
+        },
+        {
+          text: 'Empty State',
+          link: 'components/empty-state/empty-state',
+        },
+        {
+          text: 'Loader',
+          link: 'components/loader/loader',
+        },
+        {
+          text: 'Toast',
+          link: 'components/toast/toast',
+        },
+        {
+          text: 'Tooltip',
+          link: 'components/tooltip/tooltip',
+        },
+      ],
+    },
+    {
+      text: 'Data display',
+      items: [
+        {
+          text: 'Avatar',
+          link: 'components/avatar/avatar',
+        },
+        {
+          text: 'Avatar Group',
+          link: 'components/avatar-group/avatar-group',
+        },
+        {
+          text: 'Avatar Label',
+          link: 'components/avatar-label/avatar-label',
+        },
+        {
+          text: 'Badge',
+          link: 'components/badge/badge',
+        },
+        {
+          text: 'Breadcrumbs',
+          link: 'components/breadcrumbs/breadcrumbs',
+        },
+        {
+          text: 'Detail List',
+          link: 'components/detail-list/detail-list',
+        },
+        {
+          text: 'Number Badge',
+          link: 'components/number-badge/number-badge',
+        },
+        {
+          text: 'Separator',
+          link: 'components/separator/separator',
+        },
+        {
+          text: 'Text',
+          link: 'components/text/text',
+        },
+        {
+          text: 'Timeline',
+          link: 'components/timeline/timeline',
+        },
+      ],
+    },
+    {
+      text: 'Layout and navigation',
+      items: [
+        {
+          text: 'Column Layout',
+          link: 'components/column-layout/column-layout',
+        },
+        {
+          text: 'Dashboard Page',
+          link: 'components/dashboard-page/dashboard-page',
+        },
+        {
+          text: 'Row Layout',
+          link: 'components/row-layout/row-layout',
+        },
+        {
+          text: 'Sidebar',
+          link: 'components/sidebar/sidebar',
+        },
+        {
+          text: 'Tabs',
+          link: 'components/tabs/tabs',
         },
       ],
     },