+++ packages/web/design-system/src/index.ts
@@ -8,6 +8,7 @@ export {
   getLocaleFromNumberFormat, type NumberFormat,
 } from './types/numberFormat.type'
 export * from './ui'
+export { createComponent } from './utils/createComponent.util'
 export { toFormField } from './utils/toFormField.util'
 export { useNumberFormat } from '@/composables/numberFormat.composable'
 export { useIsReducedMotion } from '@/composables/useIsReducedMotion.composable'