+++ packages/web/design-system/src/ui/table/context/tableScrollContainer.context.ts
@@ -1,0 +1,8 @@
+import type { ShallowRef } from 'vue'
+
+import { useContext } from '@/composables/context.composable'
+
+export const [
+  useProvideTableScrollContainerContext,
+  useInjectTableScrollContainerContext,
+] = useContext<ShallowRef<HTMLElement | null>>('tableScrollContainerContext')