+++ packages/web/design-system/src/ui/table/index.ts
@@ -1,0 +1,58 @@
+export { default as UITable } from './components/Table.vue'
+export { default as UITableBody } from './components/TableBody.vue'
+export { default as UITableBodyGroup } from './components/TableBodyGroup.vue'
+export { default as UITableBodyRow } from './components/TableBodyRow.vue'
+export { default as UITableBodyRowActionsCell } from './components/TableBodyRowActionsCell.vue'
+export { default as UITableBodyRowCell } from './components/TableBodyRowCell.vue'
+export { default as UITableCell } from './components/TableBodyRowCell.vue'
+export { default as UITableBodyRowCellInteractiveElement } from './components/TableBodyRowCellInteractiveElement.vue'
+export { default as UITableCellInteractiveElement } from './components/TableBodyRowCellInteractiveElement.vue'
+export { default as UITableBodyRowCellText } from './components/TableBodyRowCellText.vue'
+export { default as UITableCellText } from './components/TableBodyRowCellText.vue'
+export { default as UITableBodyRowGroupCell } from './components/TableBodyRowGroupCell.vue'
+export { default as UITableBodySubGroup } from './components/TableBodySubGroup.vue'
+export { default as UITableHeader } from './components/TableHeader.vue'
+export { default as UITableHeaderCell } from './components/TableHeaderCell.vue'
+export { default as UITableHiddenResultsWarning } from './components/TableHiddenResultsWarning.vue'
+export { default as UITableLoading } from './components/TableLoading.vue'
+export { default as UITableRoot } from './components/TableRoot.vue'
+export { default as UITableScrollContainer } from './components/TableScrollContainer.vue'
+export { default as UITableSubgrid } from './components/TableSubgrid.vue'
+export { useTableColumnResize } from './composables/tableColumnResize.composable'
+export { useTableColumnWidths } from './composables/tableColumnWidths.composable'
+export { useTableFrozenColumns } from './composables/tableFrozenColumns.composable'
+export { useTableScrollState } from './composables/tableScrollState.composable'
+export { useTableVirtualScroller } from './composables/tableVirtualScroller.composable'
+export {
+  useInjectTableContext,
+  useProvideTableContext,
+} from './context/table.context'
+export {
+  useInjectTableBodyRowContext,
+  useProvideTableBodyRowContext,
+} from './context/tableBodyRow.context'
+export {
+  useInjectTableGroupContext,
+  useProvideTableGroupContext,
+} from './context/tableGroup.context'
+export {
+  useInjectTableScrollContainerContext,
+  useProvideTableScrollContainerContext,
+} from './context/tableScrollContainer.context'
+export {
+  useInjectTableSubGroupContext,
+  useProvideTableSubGroupContext,
+} from './context/tableSubGroup.context'
+export type { TableProps as UITableProps } from './types/table.props'
+export type {
+  InferTableColumnKeys as UIInferTableColumnKeys,
+  InferTableItem as UIInferTableItem,
+  TableColumn as UITableColumn,
+  TableColumnSize as UITableColumnSize,
+  TableData as UITableData,
+  TableGroupedData as UITableGroupedData,
+  TableSubGroupedData as UITableSubGroupedData,
+} from './types/table.type'
+export { defineColumns as defineTableColumns } from './types/table.type'
+export type { TableRootProps as UITableRootProps } from './types/tableRoot.props'
+export { TableUtil } from './utils/table.util'