+++ packages/api/node-doc-processor/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @wisemen/node-doc-processor
 
+## 1.3.9
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wisemen/pgboss-nestjs-job@4.0.5
+
+## 1.3.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wisemen/pgboss-nestjs-job@4.0.4
+
 ## 1.3.7
 
 ### Patch Changes